requirements.txt
# YOLOv11 Drone Detection Project Requirements
# Python 3.12+ recommended

# Core ML packages
torch>=2.8.0
torchvision>=0.23.0
ultralytics>=8.3.200

# Computer Vision
opencv-python>=4.12.0
pillow>=10.4.0

# Data Science
numpy>=2.2.0
matplotlib>=3.10.0

# Experiment tracking
wandb>=0.21.0

# Utilities
tqdm>=4.67.0
requests>=2.32.0
pyyaml>=6.0

# Optional: for additional functionality
# scipy>=1.13.0
# pandas>=2.2.0
# seaborn>=0.13.0

# Development (optional)
# pytest>=8.0.0
# black>=24.0.0
# flake8>=7.0.0